% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolloptim.R
\name{roll_max_mean}
\alias{roll_max_mean}
\title{Rolling Optimizations to Maximize Mean}
\usage{
roll_max_mean(mu, total = 1, lower = 0, upper = 1)
}
\arguments{
\item{mu}{matrix. Rows are means and columns are variables.}

\item{total}{numeric. Sum of the weights.}

\item{lower}{numeric. Lower bound of the weights.}

\item{upper}{numeric. Upper bound of the weights.}
}
\value{
An object of the same class and dimension as \code{mu} with the rolling
optimizations to maximize mean.
}
\description{
A function for computing rolling optimizations to maximize mean.
}
\examples{
if (requireNamespace("roll", quietly = TRUE)) {

n_vars <- 3
n_obs <- 15
x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)

mu <- roll::roll_mean(x, 5)

# rolling optimizations to maximize mean
roll_max_mean(mu)

}
}
