% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollport.R
\name{roll_max_return}
\alias{roll_max_return}
\title{Rolling Portfolio Optimizations to Maximize Return}
\usage{
roll_max_return(mu, sigma, total = 1, lower = 0, upper = 1)
}
\arguments{
\item{mu}{matrix. Rows are returns and columns are variables.}

\item{sigma}{cube. Slices are covariance matrices.}

\item{total}{numeric. Sum of the weights.}

\item{lower}{numeric. Lower boundry of the weights.}

\item{upper}{numeric. Upper boundry of the weights.}
}
\description{
A function for computing the rolling portfolio optimizations to maximize return.
}
