% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollport.R
\name{roll_max_utility}
\alias{roll_max_utility}
\title{Rolling Portfolio Optimizations to Maximize Utility}
\usage{
roll_max_utility(mu, sigma, gamma = 1, total = 1, lower = 0, upper = 1)
}
\arguments{
\item{mu}{matrix. Rows are returns and columns are variables.}

\item{sigma}{cube. Slices are covariance matrices.}

\item{gamma}{numeric. Risk aversion parameter.}

\item{total}{numeric. Sum of the weights.}

\item{lower}{numeric. Lower bound of the weights.}

\item{upper}{numeric. Upper bound of the weights.}
}
\value{
An object of the same class and dimension as \code{mu} with the rolling
portfolio optimizations to maximize utility.
}
\description{
A function for computing rolling portfolio optimizations to maximize utility.
}
\examples{
if (requireNamespace("roll", quietly = TRUE)) {

n_vars <- 3
n_obs <- 15
x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)

mu <- roll::roll_mean(x, 5)
sigma <- roll::roll_cov(x, width = 5)

# rolling portfolio optimizations to maximize utility
roll_max_utility(mu, sigma, gamma = 1)

}
}
