% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollport.R
\name{roll_min_risk}
\alias{roll_min_risk}
\title{Rolling Portfolio Optimization to Minimize Risks}
\usage{
roll_min_risk(mu, sigma, total = 1, lower = 0, upper = 1)
}
\arguments{
\item{mu}{matrix. Rows are returns and columns are variables.}

\item{sigma}{cube. Slices are covariance matrices.}

\item{total}{numeric. Sum of the weights.}

\item{lower}{numeric. Lower bound of the weights.}

\item{upper}{numeric. Upper bound of the weights.}
}
\description{
A function for computing rolling portfolio optimization to minimize risks.
}
