% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolloptim.R
\name{roll_min_var}
\alias{roll_min_var}
\title{Rolling Optimizations to Minimize Variance}
\usage{
roll_min_var(sigma, total = 1, lower = 0, upper = 1)
}
\arguments{
\item{sigma}{cube. Slices are covariance matrices.}

\item{total}{numeric. Sum of the weights.}

\item{lower}{numeric. Lower bound of the weights.}

\item{upper}{numeric. Upper bound of the weights.}
}
\value{
An object of the same class and dimension as \code{mu} with the rolling
optimizations to minimize variance.
}
\description{
A function for computing rolling optimizations to minimize variance.
}
\examples{
if (requireNamespace("roll", quietly = TRUE)) {

n_vars <- 3
n_obs <- 15
x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)

sigma <- roll::roll_cov(x, width = 5)

# rolling optimizations to minimize variance
roll_min_var(sigma)

}
}
