\name{rolloptim-package}
\alias{rolloptim-package}
\alias{rolloptim}
\docType{package}
\title{
  \packageTitle{rolloptim}
}
\description{
  \packageDescription{rolloptim}
}
\details{
  \code{rolloptim} is a package that provides analytical computation of rolling optimization for time-series data.
}
\author{
  \packageAuthor{rolloptim}
}
\references{
  Markowitz, H.M. (1952). "Portfolio Selection." \emph{The Journal of Finance}, 7(1), 77-91. \doi{10.1111/j.1540-6261.1952.tb01525.x}

  Tam, A. (2021). "Lagrangians and Portfolio Optimization." \url{https://www.adrian.idv.hk/2021-06-22-kkt/}
}