\name{rollport-package}
\alias{rollport-package}
\alias{rollport}
\docType{package}
\title{
\packageTitle{rollport}
}
\description{
\packageDescription{rollport}
}
\details{
\code{rollport} is a package that provides analytical computation of rolling portfolio optimization for time-series data.
}
\author{
\packageAuthor{rollport}
}
\references{
Markowitz, H.M. (1952). "Portfolio Selection." \emph{The Journal of Finance}, 7(1), 77–91.

Tam, A. (2021). "Lagrangians and Portfolio Optimization." \emph{https://www.adrian.idv.hk/2021-06-22-kkt/}.
}