% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollport.R
\name{roll_max_return}
\alias{roll_max_return}
\title{Rolling Portfolio Optimization to Maximize Returns}
\usage{
roll_max_return(mu, sigma, total = 1, lower = 0, upper = 1)
}
\arguments{
\item{mu}{matrix. Rows are returns and columns are variables.}

\item{sigma}{cube. Slices are covariance matrices.}

\item{total}{numeric. Sum of the weights.}

\item{lower}{numeric. Lower bound of the weights.}

\item{upper}{numeric. Upper bound of the weights.}
}
\description{
A function for computing rolling portfolio optimization to maximize returns.
}
\examples{
if (requireNamespace("roll", quietly = TRUE)) {

n_vars <- 3
n_obs <- 15
x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)

mu <- roll::roll_mean(x, 5)
sigma <- roll::roll_cov(x, width = 5)

# rolling portfolio optimization to maximize returns
roll_max_return(mu, sigma)

}
}
