% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollport.R
\name{roll_min_risk}
\alias{roll_min_risk}
\title{Rolling Portfolio Optimization to Minimize Risks}
\usage{
roll_min_risk(mu, sigma, total = 1, lower = 0, upper = 1)
}
\arguments{
\item{mu}{matrix. Rows are returns and columns are variables.}

\item{sigma}{cube. Slices are covariance matrices.}

\item{total}{numeric. Sum of the weights.}

\item{lower}{numeric. Lower bound of the weights.}

\item{upper}{numeric. Upper bound of the weights.}
}
\description{
A function for computing rolling portfolio optimization to minimize risks.
}
\examples{
if (requireNamespace("roll", quietly = TRUE)) {

n_vars <- 3
n_obs <- 15
x <- matrix(rnorm(n_obs * n_vars), nrow = n_obs, ncol = n_vars)

mu <- roll::roll_mean(x, 5)
sigma <- roll::roll_cov(x, width = 5)

# rolling portfolio optimization to minimize risks
roll_min_risk(mu, sigma)

}
}
