\name{rolloptim-package}
\alias{rolloptim-package}
\alias{rolloptim}
\docType{package}
\title{
  \packageTitle{rolloptim}
}
\description{
  \packageDescription{rolloptim}
}
\details{
  \code{rolloptim} is a package that provides analytical computation of rolling optimization for time-series data.
}
\author{
  \packageAuthor{rolloptim}
}
\references{
  Kuhn, H.W. and Tucker, A.W. (1951). "Nonlinear Programming." In \emph{Proceedings of the Second Berkeley Symposium on Mathematical Statistics and Probability}, edited by J. Neyman, 481-492. University of California Press. \doi{10.1525/9780520411586-036}

  Markowitz, H.M. (1952). "Portfolio Selection." \emph{The Journal of Finance}, 7(1), 77-91. \doi{10.1111/j.1540-6261.1952.tb01525.x}
}
